Data transfer bug for sensors fixed + count usable and raw energy in statistics

+++ source/EngineGpuKernels/Genome.cuh
@@ -59,8 +59,8 @@ struct SensorGenome
     uint32_t autoTriggerInterval;  // 0 = manual (triggered by signal), > 0 = auto trigger
     SensorMode mode;
     SensorModeGenome modeData;
-    int8_t minRange;  // < 0 = no restriction
-    int8_t maxRange;  // < 0 = no restriction
+    uint16_t minRange;
+    uint16_t maxRange;
 };
 
 struct ConstructorGenome
